+++ crates/protocol/src/platform.rs
@@ -20,7 +20,8 @@ use wasefire_wire::{Wire, Yoke};
 
 #[derive(Debug)]
 pub enum DynInfo {
-    V2(Yoke<Info<'static>>),
+    V3(Yoke<Info<'static>>),
+    V2(Yoke<_Info2<'static>>),
     V1(Yoke<_Info1<'static>>),
     V0(Yoke<_Info0<'static>>),
 }
@@ -29,6 +30,7 @@ pub enum DynInfo {
 impl DynInfo {
     pub fn serial(&self) -> &[u8] {
         match self {
+            DynInfo::V3(x) => &x.get().serial,
             DynInfo::V2(x) => &x.get().serial,
             DynInfo::V1(x) => &x.get().serial,
             DynInfo::V0(x) => &x.get().serial,
@@ -37,6 +39,7 @@ impl DynInfo {
 
     pub fn applet_kind(&self) -> Option<AppletKind> {
         match self {
+            DynInfo::V3(x) => Some(x.get().applet_kind),
             DynInfo::V2(x) => Some(x.get().applet_kind),
             DynInfo::V1(_) => None,
             DynInfo::V0(_) => None,
@@ -45,36 +48,114 @@ impl DynInfo {
 
     pub fn running_side(&self) -> Option<Side> {
         match self {
+            DynInfo::V3(x) => Some(x.get().running_side),
             DynInfo::V2(x) => Some(x.get().running_side),
             DynInfo::V1(x) => Some(x.get().running_side),
             DynInfo::V0(_) => None,
         }
     }
 
+    pub fn running_name(&self) -> Option<&[u8]> {
+        match self {
+            DynInfo::V3(x) => Some(&x.get().running_info.name),
+            DynInfo::V2(_) => None,
+            DynInfo::V1(_) => None,
+            DynInfo::V0(_) => None,
+        }
+    }
+
     pub fn running_version(&self) -> &[u8] {
         match self {
+            DynInfo::V3(x) => &x.get().running_info.version,
             DynInfo::V2(x) => &x.get().running_version,
             DynInfo::V1(x) => &x.get().running_version,
             DynInfo::V0(x) => &x.get().version,
         }
     }
 
-    pub fn opposite_version(&self) -> Option<Result<&[u8], Error>> {
-        let opposite = match self {
-            DynInfo::V2(x) => &x.get().opposite_version,
-            DynInfo::V1(x) => &x.get().opposite_version,
-            DynInfo::V0(_) => return None,
-        };
-        Some(match opposite {
-            Ok(x) => Ok(x),
-            Err(e) => Err(*e),
-        })
+    pub fn opposite_name(&self) -> Option<Result<&[u8], Error>> {
+        Self::opposite(
+            |x| &x[..],
+            try {
+                match self {
+                    DynInfo::V3(x) => Some(&x.get().opposite_info.as_ref()?.name),
+                    DynInfo::V2(_) => None,
+                    DynInfo::V1(_) => None,
+                    DynInfo::V0(_) => None,
+                }
+            },
+        )
     }
+
+    pub fn opposite_version(&self) -> Option<Result<&[u8], Error>> {
+        Self::opposite(
+            |x| &x[..],
+            try {
+                match self {
+                    DynInfo::V3(x) => Some(&x.get().opposite_info.as_ref()?.version),
+                    DynInfo::V2(x) => Some(x.get().opposite_version.as_ref()?),
+                    DynInfo::V1(x) => Some(x.get().opposite_version.as_ref()?),
+                    DynInfo::V0(_) => None,
+                }
+            },
+        )
+    }
+
+    fn opposite<T, R>(
+        f: impl FnOnce(T) -> R, x: Result<Option<T>, &Error>,
+    ) -> Option<Result<R, Error>> {
+        match x {
+            Ok(None) => None,
+            Ok(Some(x)) => Some(Ok(f(x))),
+            Err(e) => Some(Err(*e)),
+        }
+    }
+}
+
+/// Returns whether a platform name can be displayed as a string.
+pub fn name_str(mut name: &[u8]) -> Option<&str> {
+    loop {
+        match name.split_last() {
+            Some((0, x)) => name = x,
+            Some(_) => break,
+            None => return None,
+        }
+    }
+    name.iter().all(|x| x.is_ascii_graphic()).then(||
+        // SAFETY: We just checked that all bytes are ASCII.
+        unsafe { core::str::from_utf8_unchecked(name) })
 }
 
 #[derive(Debug, Wire)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct Info<'a> {
+    pub serial: Cow<'a, [u8]>,
+    pub applet_kind: AppletKind,
+    pub running_side: Side,
+    pub running_info: SideInfo<'a>,
+    pub opposite_info: Result<SideInfo<'a>, Error>,
+}
+
+/// Information about a platform side.
+#[derive(Debug, Wire)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+pub struct SideInfo<'a> {
+    /// Name of the platform.
+    ///
+    /// This field has not particular interpretation. However for display purposes, if the content
+    /// is only made of ASCII graphic characters after trimming the longest suffix of null bytes,
+    /// then that graphic prefix will be displayed instead of the full hexadecimal representation.
+    pub name: Cow<'a, [u8]>,
+
+    /// Version of the platform.
+    ///
+    /// This field is interpreted by lexicographical order.
+    pub version: Cow<'a, [u8]>,
+}
+
+#[derive(Debug, Wire)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+pub struct _Info2<'a> {
     pub serial: Cow<'a, [u8]>,
     pub applet_kind: AppletKind,
     pub running_side: Side,